Simulation updated; Network prioritized

+++ mcts_java/src/communication/MessageReceiver.java
@@ -5,7 +5,7 @@ import communication.messages.Message;
 public interface MessageReceiver {
     public boolean receiveQueueEmpty();
     public long receiveQueueSize();
-    public int receiveQueueLength();
+    public long receiveQueueLength();
     public Message receive();
     public Channel channel();
 }
